requirements.txt
streamlit
gTTS
moviepy
Pillow
requests
python-dotenv
tqdm
torch
torchvision
torchaudio
sentence-transformers
